:tractor: move package model gallery
