+++ lib/src/blocs/list_gallery_bloc.dart
@@ -1,6 +1,6 @@
 import 'package:rxdart/rxdart.dart';
 import '../resources/list_gallery_provider.dart';
-import '../models/list_item_gallery_model.dart';
+import '../models/gallery/list_item_gallery_model.dart';
 
 class ListGalleryBloc {
