+++ lib/src/models/gallery/list_item_gallery_model.dart
