+++ lib/src/widgets/item_gallery.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import '../data/constants.dart';
-import '../models/item_gallery_model.dart';
+import '../models/gallery/item_gallery_model.dart';
 
 class ItemGallery extends StatelessWidget {
